feat(JDK24): JEP496: Quantum-resistant module-lattice-based KEM (Kyber) - use JDK-native implementation by default
